+++ lib/embed.test.js
@@ -1,81 +1,151 @@
 'use strict';
-var embed = require('./embed'),
-  expect = require('chai').expect,
-  sinon = require('sinon'),
-  glob = require('glob'),
-  cons = require('consolidate');
+var expect = require('chai').expect,
+  proxyquire = require('proxyquire'),
+  sinon = require('sinon');
 
-describe('Embed Module', function () {
-  describe('getTemplate()', function () {
-    var sandbox;
+describe('Template embedding:', function () {
+  var fakeTemplateWithoutData = 'Fake T3mpl4te',
+    fakeTemplateData = {thing: 'stuff!'},
+    fakeMergedDefaults = {thing: 'stuff!', otherThing: 'more stuff!'},
+    expectation, embed;
 
-    beforeEach(function () {
-      sandbox = sinon.sandbox.create();
+  beforeEach(function () {
+    expectation = sinon.expectation.create('render');
+    embed = proxyquire('./embed', { './render': expectation });
+  });
+
+  describe('standalone:', function () {
+    it('Handles missing name by returning emptystring', function () {
+      var result = embed();
+      expect(result).to.equal('');
     });
 
-    afterEach(function () {
-      sandbox.restore();
+    it('Renders template with empty data', function () {
+      expectation.withArgs('jfklda', {});
+      embed('jfklda');
+      expectation.verify();
     });
 
-    it('throws error if no templates found', function () {
-      sandbox.stub(glob, 'sync').returns([]);
-      expect(function () {
-        return embed.getTemplate();
-      }).to.throw(Error);
+    it('Return data in template when piped from calling template', function () {
+      expectation.withArgs('withData', fakeTemplateData);
+      embed('withData', fakeTemplateData);
+      expectation.verify();
     });
 
-    it('throws error if templates found are not supported', function () {
-      sandbox.stub(glob, 'sync').returns(['foo.html']);
-      expect(function () {
-        return embed.getTemplate();
-      }).to.throw(Error);
+    it('Return data in template when given from extraData parameter', function () {
+      expectation.withArgs('withData', fakeTemplateData);
+      embed('withData', {}, fakeTemplateData);
+      expectation.verify();
     });
 
-    it('gets a nunjucks template', function () {
-      sandbox.stub(glob, 'sync').returns(['template.nunjucks']);
-      expect(embed.getTemplate()).to.eql('template.nunjucks');
-    });
-
-    it('gets a jade template', function () {
-      sandbox.stub(glob, 'sync').returns(['template.jade']);
-      expect(embed.getTemplate()).to.eql('template.jade');
-    });
-
-    it('gets a mustache template', function () {
-      sandbox.stub(glob, 'sync').returns(['template.mustache']);
-      expect(embed.getTemplate()).to.eql('template.mustache');
+    it('Data piped from parent has higher priority than extra parameters', function () {
+      expectation.withArgs('withData', fakeTemplateData);
+      embed('withData', fakeTemplateData, { thing: 'defaults'});
+      expectation.verify();
     });
   });
 
-  describe('render()', function () {
-    var sandbox, mock,
-    string = 'foo',
-    obj = {text: 'foo'};
+  describe('nunjucks:', function () {
+    var nunjucks = require('nunjucks'),
+      fakeTemplateWithData = 'Fake T3mpl4te {{ thing }}',
+      locals, MockLoader, env;
+
+    before(function () {
+      MockLoader = nunjucks.Loader.extend({
+        init: function () {},
+        async: false,
+        getSource: function (name) {
+          if (name.indexOf('withData') !== -1) {
+            return {src: fakeTemplateWithData};
+          } else {
+            return {src: fakeTemplateWithoutData};
+          }
+        }
+      });
+      
+      //NOTE:  Nunjucks is REALLY slow when starting up, so to prevent everyone having to wait for it whenever they run
+      // tests, we'll only be testing using a single copy.  Just a warning.
+      nunjucks.configure('views', {
+        watch: false
+      });
+      env = new nunjucks.Environment(new MockLoader());
+    });
 
     beforeEach(function () {
-      sandbox = sinon.sandbox.create();
-      mock = sinon.mock(cons);
+      locals = { embed: embed };
     });
 
-    afterEach(function () {
-      sandbox.restore();
-      mock.restore();
+    it('data passed from parent', function () {
+      var template = 'start {{ embed("withData", {thing: "stuff!"}) }}';
+
+      expectation.withArgs('withData', fakeTemplateData);
+      env.renderString(template, locals);
+      expectation.verify();
     });
 
-    it('should render nunjucks', function () {
-      sandbox.stub(glob, 'sync').returns(['foo.nunjucks']);
-      mock.expects('nunjucks').withArgs('foo.nunjucks', obj);
-      embed.render(string, obj).then(function () {
-        mock.verify();
-      });
+    it('empty passed from parent and data passed as default', function () {
+      var template = 'start {{ embed("withData", {}, {thing: "stuff!"}) }}';
+
+      expectation.withArgs('withData', fakeTemplateData);
+      env.renderString(template, locals);
+      expectation.verify();
     });
 
-    it('should render jade', function () {
-      sandbox.stub(glob, 'sync').returns(['foo.jade']);
-      mock.expects('jade').withArgs('foo.jade', obj);
-      embed.render(string, obj).then(function () {
-        mock.verify();
-      });
+    it('data from parent wins', function () {
+      var template = 'start {{ embed("withData", {thing: "stuff!"}, {thing: "not shown"}) }}';
+
+      expectation.withArgs('withData', fakeTemplateData);
+      env.renderString(template, locals);
+      expectation.verify();
+    });
+
+    it('using extraData', function () {
+      var template = 'start {{ embed("withData", {thing: "stuff!"}, {otherThing: "more stuff!"}) }}';
+
+      expectation.withArgs('withData', fakeMergedDefaults);
+      env.renderString(template, locals);
+      expectation.verify();
+    });
+  });
+
+  describe('jade function:', function () {
+    var jade = require('jade'),
+      locals;
+
+    beforeEach(function () {
+      locals = { embed: embed };
+    });
+
+    it('data passed from parent', function () {
+      var template = 'start #{embed("withData", {thing: "stuff!"})}';
+
+      expectation.withArgs('withData', fakeTemplateData);
+      jade.render(template, locals);
+      expectation.verify();
+    });
+
+    it('empty passed from parent and data passed as default', function () {
+      var template = 'start #{embed("withData", {}, {thing: "stuff!"})}';
+
+      expectation.withArgs('withData', fakeTemplateData);
+      jade.render(template, locals);
+      expectation.verify();
+    });
+
+    it('data from parent wins', function () {
+      var template = 'start #{embed("withData", {thing: "stuff!"}, {thing: "not shown"})}';
+
+      expectation.withArgs('withData', fakeTemplateData);
+      jade.render(template, locals);
+      expectation.verify();
+    });
+
+    it('using extraData', function () {
+      var template = 'start #{embed("withData", {thing: "stuff!"}, {otherThing: "more stuff!"})}';
+
+      expectation.withArgs('withData', fakeMergedDefaults);
+      jade.render(template, locals);
+      expectation.verify();
     });
   });
 });
