+++ package.json
@@ -27,6 +27,7 @@
   "devDependencies": {
     "chai": "^2.1.2",
     "mocha": "^2.2.1",
+    "proxyquire": "^1.4.0",
     "sinon": "^1.14.1"
   }
 }
